requirements.txt
Unipath==1.0
mock==1.0.1
nose==1.3.0
wsgiref==0.1.2
yanc==0.2.4
defusedxml==0.4.1
requests==2.0.0
requests-cache==0.4.3
fuzzywuzzy==0.4.0
Jinja2==2.7.3
python-Levenshtein==0.11.2
lxml==3.4.2
Django==1.9
factory-boy==2.5.2
